+++ include/logger.h
@@ -13,6 +13,14 @@ typedef enum {
     LOG_TRC
 } LOG_LEVEL;
 
+#define C_NRML "\033[0m"
+#define C_BLCK "\033[30m"
+#define C_RED  "\033[31m"
+#define C_GREN "\033[32m"
+#define C_YLLW "\033[33m"
+#define C_BLUE "\033[34m"
+#define C_PRPL "\033[35m"
+#define C_AQUA "\033[36m"
 
 #define MAX_LOCAL_TIME 64
 #define MAX_LOG_SIZE 1024
