+++ include/main.h
@@ -4,7 +4,6 @@
 #include <stdint-gcc.h>
 #include <stdio.h>
 #include <stdarg.h>
-#include <bits/types/time_t.h>
 #include <time.h>
 #include <unistd.h>
 #include <string.h>
@@ -19,12 +18,10 @@
 #include "smtpData.h"
 #include "smtpSession.h"
 #include "smtpSock.h"
-#include "smtpSvrRecvFdSet.h"
 #include "smtpUtils.h"
 #include "smtpDispatch.h"
 #include "smtpMailFile.h"
-#include "smtpSvrRecvEpoll.h"
-#include "smtpItcq.h"
+#include "smtpSvrRecvAsync.h"
 
 extern int g_sys_close;
