--- include/smtpItcq.h
@@ -1,6 +1,0 @@
-#ifndef CAREER_TEST3_SMTPITCQ_H
-#define CAREER_TEST3_SMTPITCQ_H
-extern smtp_session_t * itcqGetSession();
-extern int itcqPutSession(smtp_session_t * session);
-
-#endif //CAREER_TEST3_SMTPITCQ_H
