+++ include/smtpSvrRecvAsync.h
@@ -1,0 +1,5 @@
+#ifndef CAREER_TEST3_SMTPSVRRECVASYNC_H
+#define CAREER_TEST3_SMTPSVRRECVASYNC_H
+extern void smtpWaitAsync(int server_fd);
+extern int smtpStartWorkThreads(int n_work_threads);
+#endif //CAREER_TEST3_SMTPSVRRECVASYNC_H
