--- include/smtpSvrRecvEpoll.h
@@ -1,5 +1,0 @@
-#ifndef CAREER_TEST3_SMTPSVRRECVEPOLL_H
-#define CAREER_TEST3_SMTPSVRRECVEPOLL_H
-extern void smtpWaitAsync(int server_fd);
-extern int smtpStartWorkThreads(int n_work_threads);
-#endif //CAREER_TEST3_SMTPSVRRECVEPOLL_H
