--- include/smtpSvrRecvFdSet.h
@@ -1,5 +1,0 @@
-#ifndef CAREER_TEST3_SMTPSVRRECVFDSET_H
-#define CAREER_TEST3_SMTPSVRRECVFDSET_H
-extern void * H_SERVER_RECV_FDSET_TH ( void * args );
-extern void smtpWaitSync(int server_fd);
-#endif //CAREER_TEST3_SMTPSVRRECVFDSET_H
